resnet_classification_movie_type/requirements.txt
torchvision==0.4.0
matplotlib==3.3.2
pandas==1.1.2
numpy==1.19.2
torch==1.2.0
asposestorage==1.0.2
Pillow==8.1.0
